refactor: クラスの可視性を修正
